requirements.txt
nose
zope.component
